added pagination to correct position + functions for handling the page number
